Fix failing :on_replace_update tests
This fixes the failing :on_replace_update tests. They were throwing the following error:
```
(Ecto.StaleEntryError) attempted to update a stale struct
```

The underlying Mongo query being issued returned a document with `nModified` set to `0`. Ecto's changeset typically detects a no-op change but doesn't in this scenario. If Mongo matched the document (even if it didn't modify it), then a stale error should not be thrown.

Based off of git history blames, here is the timeline of events that show why this bug is only now appearing:
1. `mongodb_ecto` depends on mongodb v0.4.x
2. `mongodb` v0.5.x changed the logic (see here: https://github.com/elixir-mongo/mongodb/commit/7f6181e121a83f75e983655ec22e2f454347d31e). It’s actually a bug fix because now it is setting it to the correct nModified
3. The `mongodb`  Ecto 3 branch built off of `mongodb` v0.5.x
4. The bug now exists